refactor: moved camera logic and permission from main to result activity
