+++ app/src/main/java/com/rempahpedia/rempahpedia/ui/classification/ResultActivity.kt
@@ -1,25 +1,43 @@
 package com.rempahpedia.rempahpedia.ui.classification
 
+import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import com.rempahpedia.rempahpedia.R
+import androidx.core.net.toUri
 import com.rempahpedia.rempahpedia.databinding.ActivityResultBinding
+import com.rempahpedia.rempahpedia.ui.classification.CameraActivity.Companion.CAMERAX_RESULT
 
 class ResultActivity : AppCompatActivity() {
     private lateinit var binding: ActivityResultBinding
+    private var currentImageUri: Uri? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        startCameraX()
+
         super.onCreate(savedInstanceState)
         binding = ActivityResultBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        val imageUri = Uri.parse(intent.getStringExtra(IMAGE_URI))
-        binding.resultImage.setImageURI(imageUri)
-        binding.resultText.text = getString(R.string.ml_soon)
     }
 
-    companion object {
-        const val IMAGE_URI = "imageURI"
+    private fun showImage() {
+        currentImageUri?.let {
+            binding.resultImage.setImageURI(it)
+        }
+    }
+
+    private fun startCameraX() {
+        val intent = Intent(this@ResultActivity, CameraActivity::class.java)
+        launcherIntentCameraX.launch(intent)
+    }
+
+    private val launcherIntentCameraX = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) {
+        if (it.resultCode == CAMERAX_RESULT) {
+            currentImageUri = it.data?.getStringExtra(CameraActivity.EXTRA_CAMERAX_IMAGE)?.toUri()
+            showImage()
+        }
     }
 }
